# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Amafor Gladiators Digital Ecosystem API
  description: |
    Complete REST API specification with full traceability to SRS requirements.
    Each endpoint includes custom extension `x-srs-req` for requirement mapping.
  version: 1.0.0
  contact:
    name: Engineering Team
    email: dev@amaforgladiatorsfc.com
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.amaforgladiatorsfc.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /articles/tag/{tag}:
    get:
      tags: [Articles]
      summary: Get articles by tag
      operationId: API-ARTICLE-003
      x-srs-req: REQ-PUB-03
      parameters:
        - name: tag
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Articles by tag
  /articles/search:
    get:
      tags: [Articles]
      summary: Search articles
      operationId: API-ARTICLE-004
      x-srs-req: REQ-PUB-03
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results
...
  /articles/{id}:
    get:
      tags: [Articles]
      summary: Get article by ID
      operationId: API-ARTICLE-006
      x-srs-req: [REQ-PUB-03, REQ-CMS-01]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article details

  # League Statistics
...